# docs/swagger.yaml
basePath: /
definitions:
  models.CreateSubscriptionRequest:
    properties:
      end_date:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  models.SubscriptionResponse:
    properties:
      end_date:
        type: string
      id:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  models.TotalCostResponse:
    properties:
      total_cost:
        type: integer
    type: object
  models.UpdateSubscriptionRequest:
    properties:
      end_date:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST сервис для управления подписками пользователя.
  title: Subscriptions API
  version: "1.0"
paths:
  /subscriptions:
    get:
      parameters:
      - description: User UUID (optional - returns all if not provided)
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SubscriptionResponse'
            type: array
        "400":
          description: Invalid parameters
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Получить список подписок
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Создание новой подписки. Поле `end_date` опциональное. Если не
        указано - подписка бессрочная.
      parameters:
      - description: Subscription info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SubscriptionResponse'
        "400":
          description: Invalid request body
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Создать подписку
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid ID format
          schema:
            type: string
        "404":
          description: Subscription not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Удалить подписку
      tags:
      - subscriptions
    get:
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SubscriptionResponse'
        "400":
          description: Invalid ID format
          schema:
            type: string
        "404":
          description: Subscription not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Получить подписку по ID
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Обновление данных подписки. Все поля опциональные. При обновлении
        `end_date` проверяется, что `end_date >= start_date`.
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      - description: Subscription update info (all fields optional)
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.UpdateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SubscriptionResponse'
        "400":
          description: Invalid request
          schema:
            type: string
        "404":
          description: Subscription not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Обновить подписку
      tags:
      - subscriptions
  /subscriptions/total:
    get:
      description: |-
        Расчёт общей стоимости активных подписок за указанный период.<br><br>
        **Логика расчёта:**<br>
        - Подписка с `end_date` учитывается, если пересекается с запрошенным периодом (start_date <= end_period AND end_date >= start_period)<br>
        - Подписка без `end_date` (бессрочная) учитывается полностью, если её `start_date <= end_period`<br>
        - Бессрочная подписка всегда учитывается полной стоимостью, независимо от длины запрошенного периода
      parameters:
      - description: User UUID
        in: query
        name: user_id
        required: true
        type: string
      - description: 'Format: MM-YYYY'
        in: query
        name: start_date
        required: true
        type: string
      - description: 'Format: MM-YYYY'
        in: query
        name: end_date
        required: true
        type: string
      - description: Service Name filter (optional)
        in: query
        name: service_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TotalCostResponse'
        "400":
          description: Invalid parameters
          schema:
            type: string
        "404":
          description: No subscriptions found for the specified criteria
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Получение общей стоимости подписок за заданный период
      tags:
      - subscriptions
swagger: "2.0"